+++ kaithem/src/docs/changes.md
@@ -4,8 +4,11 @@ Change Log
 
 ### 0.89.0
 
+This release has one of the first AI features! We use sherpa-onnx and and automatically download models.
+
 - :sparkles: New system tag points and default alarms for hdmi port connection(/system/display_ports/hdmi_1.connected)
 - :sparkles: Stateful blocks like on_change in the visual logic editor.
+- :sparkles: Speak command that useds the new globally configured TTS model.
 - :bug: Fix web media not rendering without login
 - :bug: Fix mixer sort issue and use fixed width num param fields